pyproject.toml
[project]
name = "uiautoagent"
version = "0.1.0"
description = "AI-powered UI element detection and device automation framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "json-repair>=0.59.4",
    "markdown>=3.10.2",
    "numpy>=1.24.0",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.10.0",
    "wdapy>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.scripts]
uiautoagent = "uiautoagent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uiautoagent"]

[dependency-groups]
dev = [
    "pytest-mock>=3.15.1",
    "ruff>=0.15.10",
    "ruff-lsp>=0.0.62",
]
